Refactor Pages
