+++ src/FamilyHub.IdentityServerHost/Areas/Gds/Pages/Invitation/Confirmation.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.Manage.ConfirmationModel
+@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.Invitation.ConfirmationModel
 @{
 }
