+++ src/FamilyHub.IdentityServerHost/Areas/Gds/Pages/Invitation/WhatIsEmailAddress.cshtml
@@ -1,9 +1,9 @@
 ﻿@page
-@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.Manage.WhatIsEmailAddressModel
+@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.Invitation.WhatIsEmailAddressModel
 @{
 }
 
-<a asp-page="/Manage/WhatIsUsername"  class="govuk-back-link" data-testid="back-button">Back</a>
+<a asp-page="/Invitation/WhatIsUsername"  class="govuk-back-link" data-testid="back-button">Back</a>
 
 
 <main class="govuk-main-wrapper govuk-main-wrapper--auto-spacing" id="main-content" role="main">
