+++ src/FamilyHub.IdentityServerHost/Areas/Gds/Pages/Invitation/WhichOrganisation.cshtml
@@ -1,9 +1,9 @@
 ﻿@page
-@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.Manage.WhichOrganisationModel
+@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.Invitation.WhichOrganisationModel
 @{
 }
 
-<a asp-page="/Manage/TypeOfUser" class="govuk-back-link" data-testid="back-button">Back</a>
+<a asp-page="/Invitation/TypeOfUser" class="govuk-back-link" data-testid="back-button">Back</a>
 
 <main class="govuk-main-wrapper govuk-main-wrapper--auto-spacing" id="main-content" role="main">
     <div>
