+++ src/FamilyHub.IdentityServerHost/Areas/Gds/Pages/Manage/Homepage.cshtml
@@ -13,7 +13,7 @@
 <hr class="govuk-section-break govuk-section-break--visible govuk-section-break--l">
 <div class="cards">
     <div class="card">
-        <h2 class="govuk-heading-m"><a href="TypeOfUser">Create user accounts</a></h2>
+        <h2 class="govuk-heading-m"><a asp-page="/Invitation/TypeOfUser">Create user accounts</a></h2>
         <p class="govuk-body-s">Create different types of user accounts.</p>
     </div>
     <div class="card">
@@ -21,7 +21,7 @@
         <p class="govuk-body-s">View, change or delete user accounts.</p>
     </div>
     <div class="card">
-        <h2 class="govuk-heading-m"><a href="TypeOfOrganisation">Add a local authority or voluntary community organisation</a></h2>
+            <h2 class="govuk-heading-m"><a asp-page="/OrganisationWizard/TypeOfOrganisation">Add a local authority or voluntary community organisation</a></h2>
         <p class="govuk-body-s"> Add a local authority or organisation before creating their user accounts.</p>
     </div>
     <div class="card">
