+++ src/FamilyHub.IdentityServerHost/Areas/Gds/Pages/OrganisationWizard/OrganisationName.cshtml
@@ -1,9 +1,9 @@
 ﻿@page
-@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.Manage.OrganisationNameModel
+@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.OrganisationWizard.OrganisationNameModel
 @{
 }
 
-<a asp-page="/Manage/WhichLAOrAdminDistrict" class="govuk-back-link" data-testid="back-button">Back</a>
+<a asp-page="/OrganisationWizard/WhichLAOrAdminDistrict" class="govuk-back-link" data-testid="back-button">Back</a>
 
 <main class="govuk-main-wrapper govuk-main-wrapper--auto-spacing" id="main-content" role="main">
