+++ src/FamilyHub.IdentityServerHost/Areas/Gds/Pages/OrganisationWizard/TypeOfOrganisation.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.Manage.TypeOfOrganisationModel
+@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.OrganisationWizard.TypeOfOrganisationModel
 @{
 }
