+++ src/FamilyHub.IdentityServerHost/Areas/Gds/Pages/OrganisationWizard/WhichLAOrAdminDistrict.cshtml
@@ -1,9 +1,9 @@
 ﻿@page
-@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.Manage.WhichLAOrAdminDistrictModel
+@model FamilyHub.IdentityServerHost.Areas.Gds.Pages.OrganisationWizard.WhichLAOrAdminDistrictModel
 @{
 }
 
-<a asp-page="/Manage/TypeOfOrganisation" class="govuk-back-link" data-testid="back-button">Back</a>
+<a asp-page="/OrganisationWizard/TypeOfOrganisation" class="govuk-back-link" data-testid="back-button">Back</a>
 
 <main class="govuk-main-wrapper govuk-main-wrapper--auto-spacing" id="main-content" role="main">
      <fieldset class="govuk-fieldset">
